[chore] GithubDetailAdapter -> GithubAdapter로 클래스명 수정 및 inner class -> class로 분리
